requirements.txt
Flask==3.0.3
pymongo==4.8.0
requests==2.32.3
python-dotenv==1.0.1
flasgger==0.9.7b2
gunicorn==22.0.0
Jinja2==3.1.4
